separate MigrationVrmMeta
