+++ Assets/VRM10/Runtime/Migration/MigrationVrmMeta.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: bb019cd042148c2428f4e48b9391c448
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
